+++ .gitignore
@@ -1,4 +1,5 @@
 bentchmark.ipynb
+test_gpu.py
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
