+++ changelog.md
@@ -116,8 +116,10 @@
 
 ### 0.0.3 (7/4/2025)
 
-- **Redesigned `GLpoint`:** Now uses a fast, single-pass recurrence to compute the Grünwald-Letnikov fractional derivative at the endpoint. This new implementation is both efficient and direct.
-- Legacy and alternative endpoint methods (`GLpoint_direct` and `GLpoint_via_GL`) have been moved to `special.py` for reference and comparison.
+* **Redesigned `GLpoint`:**
+  Replaced the previous GLpoint implementation with a highly efficient, single-pass C++-style recurrence. The new method uses a Numba JIT-compiled kernel for substantial speed gains, directly computing the Grünwald-Letnikov fractional derivative at the endpoint in one loop over the function values.
+  Legacy and alternative endpoint methods (`GLpoint_direct`, `GLpoint_via_GL`) have been moved to `special.py` for reference and comparison.
+
 
 [Benchmark](https://github.com/iparsw/differintC/blob/main/BENCHMARK.md) Result for GLpoint:
 
@@ -131,23 +133,27 @@
 | 1+e7  | 51.2645 ms  | 3674.3541 ms | x71.67         |
 
 
-- **Optimized `GLI` implementation:**
+* **Optimized `GLI` implementation:**
 
-  - Replaced class-based coefficient calculation with direct variable computation for improved clarity and performance.
-  - Vectorized the 3-point Lagrange interpolation step for all interior points.
-  - Switched from per-point convolution in a loop to a single global convolution for the entire array, greatly increasing speed.
-  - Introduced a hybrid approach: uses direct convolution for arrays smaller than 800 points and FFT convolution for larger arrays to ensure optimal performance.
+  * Rewrote core algorithm to use a Numba-accelerated helper for the main rolling-window convolution, matching the original literature.
+  * Ensured correctness by flipping GL coefficients within the moving-window convolution.
+  * Replaced class-based coefficient calculation with direct variable computation for improved clarity and efficiency.
+  * All critical loops are now compiled with Numba for major speed improvements, especially at high grid resolutions.
+  * Modernized function interface and added robust array conversion and shape checks.
 
 [Benchmark](https://github.com/iparsw/differintC/blob/main/BENCHMARK.md) Result for GLI:
 
 | count |  differintP | differint    | Speedup Factor |
 | ----- | ----------- | ------------ | -------------- |
-| 1+e2  | 0.7582 ms   | 0.0536 ms    | x14.14         |
-| 1+e3  | 7.391 ms    | 0.1832 ms    | x40.34         |
-| 1+e4  | 128.3709 ms | 0.6801 ms    | x188.7         |
-| 2+e4  | 15181  ms   | 5.3272 ms    | x2849          |
+| 1+e2  | 0.0760 ms   | 0.7582 ms    | x6.06          |
+| 1+e3  | 2.0336 ms   | 7.391 ms     | x3.63          |
+| 1+e4  | 155.7649 ms | 128.3709 ms  | x0.8           |
+| 5+e4  | 3895.176 ms | 15181.82 ms  | x3.89          |
 
 - Wiki Pages for
   - `CaputoL1point` `CaputoL2point` `CaputoL2Cpoint` `CaputoFromRLpoint`
   - `CRONE` `GL` `GLpoint` `GLI` `GL_gpu`
-  - `RL` `RLpoint`
+  - `RL` `RLpoint` `PCsolver` functions Namespaces
+
+- Aditional test and test restructuring
+- Restructuring the Namespaces
