+++ differintP/__init__.py
@@ -1,15 +1,37 @@
-from .core import *
+from .core import (
+    GLcoeffs,
+    GL,
+    GLpoint,
+    GL_gpu,
+    GLI,
+    CRONE,
+    RLmatrix,
+    RLcoeffs,
+    RLpoint,
+    RL,
+    CaputoL1point,
+    CaputoL2point,
+    CaputoL2Cpoint,
+    CaputoFromRLpoint,
+)
+
+from . import functions
+
 __all__ = [
-    "GL",
     "GLcoeffs",
-    "functionCheck",
-    "checkValues",
-    "isInteger",
-    "GLI",
+    "GL",
     "GLpoint",
-    "RLcoeffs",
+    "GL_gpu",
+    "GLI",
+    "CRONE",
     "RLmatrix",
-    "RL",
-    "poch",
+    "RLcoeffs",
     "RLpoint",
+    "RL",
+    "CaputoL1point",
+    "CaputoL2point",
+    "CaputoL2Cpoint",
+    "CaputoFromRLpoint",
+
+    "functions"
 ]
