+++ tests/__init__.py
@@ -1,0 +1,1 @@
+test_N = 512
